+++ program/system/global.h
@@ -67,7 +67,7 @@ typedef struct {
 
 global_t variable[SYS_VAR_CNT];
 
-extern uint64_t boot_time;
+extern uint32_t boot_time;
 
 extern system_t system;
